feat: Refactor AtomicOp.Do
