+++ internal/pkg/service/common/etcdop/op/atomic.go
@@ -2,10 +2,6 @@ package op
 
 import (
 	"context"
-	"github.com/cenkalti/backoff/v4"
-	"github.com/keboola/keboola-as-code/internal/pkg/utils/errors"
-	"go.etcd.io/etcd/api/v3/etcdserverpb"
-	"time"
 
 	etcd "go.etcd.io/etcd/client/v3"
 )
@@ -167,132 +163,4 @@ func (v *AtomicOp[R]) OnResult(fn func(result R)) *AtomicOp[R] {
 func (v *AtomicOp[R]) EmptyResultAsError(fn func() error) *AtomicOp[R] {
 	v.processors = v.processors.WithEmptyResultAsError(fn)
 	return v
-}
-
-func (v *AtomicOp[R]) Do(ctx context.Context, opts ...Option) AtomicResult[R] {
-	b := newBackoff(opts...)
-	attempt := 0
-
-	var ok bool
-	var err error
-
-	for {
-		txnResult := v.DoWithoutRetry(ctx, opts...)
-		ok = txnResult.Succeeded()
-		err = txnResult.Err()
-		if err == nil && !ok {
-			attempt++
-			if delay := b.NextBackOff(); delay == backoff.Stop {
-				break
-			} else {
-				<-time.After(delay)
-			}
-		} else {
-			break
-		}
-	}
-
-	elapsedTime := b.GetElapsedTime()
-	if err == nil && !ok {
-		err = errors.Errorf(
-			`atomic update failed: revision has been modified between GET and UPDATE op, attempt %d, elapsed time %s`,
-			attempt, elapsedTime,
-		)
-	}
-
-	return AtomicResult[R]{result: v.result, error: err, attempt: attempt, elapsedTime: elapsedTime}
-}
-
-func (v *AtomicOp[R]) DoWithoutRetry(ctx context.Context, opts ...Option) *TxnResult[R] {
-	tracker := NewTracker(v.client)
-
-	// Create READ operations
-	readTxn, err := v.readTxn(ctx, tracker)
-	if err != nil {
-		return newErrorTxnResult[R](err)
-	}
-
-	// Run READ phase, track used keys/prefixes
-	readResult := readTxn.Do(ctx, opts...)
-	if err := readResult.Err(); err != nil {
-		return newErrorTxnResult[R](err)
-	}
-
-	// Create WRITE transaction
-	writeTxn, err := v.writeTxn(ctx)
-	if err != nil {
-		return newErrorTxnResult[R](err)
-	}
-
-	// Add IF conditions
-	writeTxn.If(v.writeIfConditions(tracker, readResult.Header().Revision)...)
-
-	// Do
-	return writeTxn.Do(ctx)
-}
-
-
-// x Create IF part of the transaction.
-func (v *AtomicOp[R]) writeIfConditions(tracker *TrackerKV, readRev int64) (cmps []etcd.Cmp) {
-	for _, op := range tracker.Operations() {
-		mustExist := (op.Type == GetOp || op.Type == PutOp) && op.Count > 0
-		mustNotExist := op.Type == DeleteOp || op.Count == 0
-		switch {
-		case mustExist:
-			// IF: 0 < modification version <= Read Phase revision
-			// Key/range exists and has not been modified since the Read Phase.
-			//
-			// Note: we cannot check that nothing was deleted from the prefix.
-			// The condition IF count == n is needed, and it is not implemented in etcd.
-			// We can verify that an individual key was deleted, its MOD == 0.
-			cmps = append(cmps,
-				// The key/prefix must exist, version must be NOT equal to 0.
-				etcd.Cmp{
-					Target:      etcdserverpb.Compare_MOD,
-					Result:      etcdserverpb.Compare_GREATER,
-					TargetUnion: &etcdserverpb.Compare_ModRevision{ModRevision: 0},
-					Key:         op.Key,
-					RangeEnd:    op.RangeEnd, // may be empty
-				},
-				// The key/prefix cannot be modified between GET and UPDATE phase.
-				// Mod revision of the item must be less or equal to header.Revision.
-				etcd.Cmp{
-					Target:      etcdserverpb.Compare_MOD,
-					Result:      etcdserverpb.Compare_LESS, // see +1 bellow, so less or equal to header.Revision
-					TargetUnion: &etcdserverpb.Compare_ModRevision{ModRevision: readRev + 1},
-					Key:         op.Key,
-					RangeEnd:    op.RangeEnd, // may be empty
-				},
-			)
-
-			// See SkipPrefixKeysCheck method documentation, by default, the feature is enabled.
-			if v.checkPrefixKey {
-				if op.RangeEnd != nil {
-					for _, kv := range op.KVs {
-						cmps = append(cmps, etcd.Cmp{
-							Target:      etcdserverpb.Compare_MOD,
-							Result:      etcdserverpb.Compare_GREATER,
-							TargetUnion: &etcdserverpb.Compare_ModRevision{ModRevision: 0},
-							Key:         kv.Key,
-						})
-					}
-				}
-			}
-		case mustNotExist:
-			cmps = append(cmps,
-				// IF: modification version == 0
-				// The key/range doesn't exist.
-				etcd.Cmp{
-					Target:      etcdserverpb.Compare_MOD,
-					Result:      etcdserverpb.Compare_EQUAL,
-					TargetUnion: &etcdserverpb.Compare_ModRevision{ModRevision: 0},
-					Key:         op.Key,
-					RangeEnd:    op.RangeEnd, // may be empty
-				},
-			)
-		default:
-			panic(errors.Errorf(`unexpected state, operation type "%v"`, op.Type))
-		}
-	}
-	return cmps
 }
